Add abstract BaseParquetReaders for Iceberg generics and Flink (#1162)
