+++ flink/src/main/java/org/apache/iceberg/flink/data/FlinkParquetWriters.java
@@ -21,31 +21,25 @@ package org.apache.iceberg.flink.data;
 
 import java.util.List;
 import org.apache.flink.types.Row;
-import org.apache.iceberg.data.parquet.GenericParquetWriter;
-import org.apache.iceberg.parquet.ParquetTypeVisitor;
+import org.apache.iceberg.data.parquet.BaseParquetWriter;
 import org.apache.iceberg.parquet.ParquetValueWriter;
 import org.apache.iceberg.parquet.ParquetValueWriters;
 import org.apache.parquet.schema.MessageType;
 
-public class FlinkParquetWriters extends GenericParquetWriter {
+public class FlinkParquetWriters extends BaseParquetWriter<Row> {
+
+  private static final FlinkParquetWriters INSTANCE = new FlinkParquetWriters();
+
   private FlinkParquetWriters() {
   }
 
-  @SuppressWarnings("unchecked")
-  public static ParquetValueWriter<Row> buildRowWriter(MessageType type) {
-    return (ParquetValueWriter<Row>) ParquetTypeVisitor.visit(type, new WriteBuilder(type));
+  public static ParquetValueWriter<Row> buildWriter(MessageType type) {
+    return INSTANCE.createWriter(type);
   }
 
-  private static class WriteBuilder extends GenericParquetWriter.WriteBuilder {
-
-    private WriteBuilder(MessageType type) {
-      super(type);
-    }
-
-    @Override
-    protected ParquetValueWriters.StructWriter<Row> createStructWriter(List<ParquetValueWriter<?>> writers) {
-      return new RowWriter(writers);
-    }
+  @Override
+  protected ParquetValueWriters.StructWriter<Row> createStructWriter(List<ParquetValueWriter<?>> writers) {
+    return new RowWriter(writers);
   }
 
   private static class RowWriter extends ParquetValueWriters.StructWriter<Row> {
